feat: friendly thug

+++ Dal/Entities/Project.cs
@@ -2,10 +2,10 @@
 
 namespace Dal.Entities;
 
-public class Project : TrackableEntity<int>
+public class Project : TrackableEntity<long>
 {
     public string Name { get; set; }
     
-    public int OrganizationId { get; set; }
+    public long OrganizationId { get; set; }
     public Organization Organization { get; set; }
 }
