+++ Dal/Entities/TaskCollaborator.cs
@@ -2,11 +2,11 @@
 
 namespace Dal.Entities;
 
-public class TaskCollaborator : BaseEntity<int>
+public class TaskCollaborator : BaseEntity<long>
 {
-    public int TaskId { get; set; }
+    public long TaskId { get; set; }
     public Task Task { get; set; }
 
-    public int CollaboratorId { get; set; }
+    public long CollaboratorId { get; set; }
     public User Collaborator { get; set; }
 }
